Moved current grid layout code into its own file
